utils: Move the file archiver to the `model` package
The file archiver will be extended to support a PostgreSQL based storage
in addition to the file based storage. Thus, the file archiver's API
must be changed to abstract from the storage path, because that path
is specific to the file storage and an implementation detail.

This in turn implies that the file archiver has to derive the storage
path itself from `identifier` and `provenance`. Move the file archiver
to model module to enable using these two mentioned (model) classes.
